airflow/requirements.txt
apache-airflow==2.8.1
snowflake-connector-python==3.6.0
requests==2.31.0
pandas==2.1.4
psycopg2-binary==2.9.9
